Extracted classes to separate files.
